video-service/requirements.txt
django
boto3
djangorestframework